Changes to support docfx
Note: This requires a patch on the docfx executable, the hashbang line is
missing, creating an execution failure

This makes the executed command docfx, instead of BuildMeta. It adds some error
checking and reporting to output as well.

This is a work in progress, the output format seems to have changed again.