Support specifying package directories as well
At the moment, if you build the docs in this repo, the autoapi docs are rooted by their subpackages, i.e. listed as `mappers` instead of `autoapi.mappers`. The latter is the one that's expected, as when you import, you do `import autoapi.mappers`. The incorrect module path permeates through all the docs.

This change fixes that by changing the `autoapi_dirs` config parameter - if a directory specified has a `__init__.py` under it, the directory itself will also be included in the relative paths, resulting in `autoapi.mappers`. The previous behaviour is still there though, as if the directory does **not** have a `__init__.py` under it, the children will be documented with their path as the root, i.e. `mappers`.

I've added tests to demonstrate this.

I believe this addresses #108.