Use os.sep instead of the hardcoded '/'
This was making the tests fail on Windows because the necessary '/'
replacements were not happening on Windows due to os.sep being '\\'.

I've also updated a test so it uses whatever the directory separator is on
the system. Ideally the whole code base would use os.sep instead, but this
makes all the tests pass on Windows.