Fix Go documentation generation (#172)
* Correction of the "parameters" accessibility problem in the Go mapper

The "parameters" attributes of the GoPythonMapper class instance was
not accessible from the "output_rst" function used to create the .rst
file. Thanks to this correction, the parameters are now accessible.

* Enhance the template used to genertae the .rst file from the .go source file

This modification allows to create a .rst file compatible with the
last version of Sphinx. Without this correction, the creation of documentation
was not possible. A "make html" command lead to an "AssertionError" Exception.